move auth logic
